ICMSLST-413 Reduce query number on importer
Prefetch `offices` as they are displayed in every row of the `importer-list`
view. Prefetching offices avoid to issue a query every time an office should
be displayed.
See https://docs.djangoproject.com/en/2.2/ref/models/querysets/#django.db.models.query.QuerySet.prefetch_related.
The prefetch reduces the queries from 53 (1 office for each row) to 4.

Fix `select_related` which was targeting the wrong model. It was `user`
and the queryset is filtering on `main_importer`.

Prefetch `offices` as they are displayed in every row of the
`importer-list` view. And fix `JOIN` on `main_importer`.

`importer-list` now uses now 4 queries instead of 53.